requirements.txt
jupyterlab
ipykernel
pandas
matplotlib
scikit-learn
keras
Pillow
colorspacious
numpy
opencv-python
json5
keras
tensorflow